Move Storage service out of analytics and into core
